feat: Migrate to @posthog/mcp-analytics completely
We migrated behind a FF but the code is getting too complicated, so let's just migrate fully, the FF is already at 100% in production

Also use v0.0.9 to get some more niceties for `get_more_tools` and the `identity` field being a non-function

Also moved some stuff around to better reflect this is all analytics
