+++ services/mcp/package.json
@@ -41,7 +41,7 @@
         "@hono/node-server": "^2.0.2",
         "@modelcontextprotocol/ext-apps": "^1.5.0",
         "@modelcontextprotocol/sdk": "^1.29.0",
-        "@posthog/mcp-analytics": "npm:@posthog/mcp@0.0.7",
+        "@posthog/mcp-analytics": "npm:@posthog/mcp@0.0.9",
         "@posthog/quill": "workspace:*",
         "@toon-format/toon": "^2.1.0",
         "agents": "0.10.2",
@@ -52,7 +52,6 @@
         "hono": "^4.12.18",
         "ioredis": "^4.28.5",
         "lucide-react": "^0.577.0",
-        "mcpcat": "0.1.15",
         "posthog-js-lite": "4.2.2",
         "posthog-node": "^5.25.0",
         "prom-client": "^14.2.0",
