+++ services/mcp/src/lib/posthog/flags.ts
@@ -1,0 +1,30 @@
+import { getPostHogClient } from './analytics'
+
+export async function isFeatureFlagEnabled(flagKey: string, distinctId: string): Promise<boolean> {
+    try {
+        const client = getPostHogClient()
+        const result = await client.isFeatureEnabled(flagKey, distinctId)
+        return result === true
+    } catch {
+        return false
+    }
+}
+
+/**
+ * Evaluate multiple feature flags in parallel for the given user.
+ * Returns a map of flag key → boolean.
+ */
+export async function evaluateFeatureFlags(flagKeys: string[], distinctId: string): Promise<Record<string, boolean>> {
+    if (flagKeys.length === 0) {
+        return {}
+    }
+
+    const results = await Promise.all(
+        flagKeys.map(async (key) => {
+            const enabled = await isFeatureFlagEnabled(key, distinctId)
+            return [key, enabled] as const
+        })
+    )
+
+    return Object.fromEntries(results)
+}
