+++ services/mcp/src/mcp.ts
@@ -5,15 +5,6 @@ import { McpAgent } from 'agents/mcp'
 import type { z } from 'zod'
 
 import { ApiClient } from '@/api/client'
-import {
-    AnalyticsEvent,
-    buildMCPAnalyticsGroups,
-    buildMCPContextProperties,
-    evaluateFeatureFlags,
-    getPostHogClient,
-    isFeatureFlagEnabled,
-    type MCPAnalyticsContext,
-} from '@/lib/analytics'
 import { hasScope } from '@/lib/api'
 import { buildToolResultPayload, isToolCallPayload } from '@/lib/build-tool-result'
 import { DurableObjectCache } from '@/lib/cache/DurableObjectCache'
@@ -28,8 +19,15 @@ import {
 import { handleToolError, wrapError } from '@/lib/errors'
 import { type QueryToolInfo } from '@/lib/instructions'
 import { InstructionsFormatter } from '@/lib/instructions-formatter'
-import { initMcpCatObservability } from '@/lib/mcpcat'
-import { initPostHogMcpAnalytics, type PostHogMcpAnalyticsInitResult } from '@/lib/posthog-mcp-analytics'
+import {
+    AnalyticsEvent,
+    buildMCPAnalyticsGroups,
+    buildMCPContextProperties,
+    getPostHogClient,
+    initMcpAnalytics,
+    McpAnalyticsInitResult,
+    type MCPAnalyticsContext,
+} from '@/lib/posthog/analytics'
 import { SessionManager } from '@/lib/SessionManager'
 import { StateManager } from '@/lib/StateManager'
 import { formatPrompt, type McpMode, sanitizeHeaderValue } from '@/lib/utils'
@@ -41,14 +39,9 @@ import { createExecInnerToolCallResolver, createExecTool, type ExecInnerCallTrac
 import { getToolDefinition } from '@/tools/toolDefinitions'
 import { type CloudRegion, type Context, type State, type Tool } from '@/tools/types'
 
+import { evaluateFeatureFlags, isFeatureFlagEnabled } from './lib/posthog/flags'
+
 const instructionsFormatter = new InstructionsFormatter()
-const POSTHOG_MCP_ANALYTICS_FLAG = 'mcp-posthog-analytics-sdk'
-type McpAnalyticsProvider = 'posthog_mcp_analytics' | 'mcpcat'
-type PostHogMcpAnalyticsFlagResult = {
-    enabled: boolean
-    errorName?: string
-    errorMessage?: string
-}
 
 export type RequestProperties = {
     userHash: string
@@ -86,15 +79,10 @@ export type RequestProperties = {
     transport?: 'streamable-http' | 'sse'
     viaSseRedirect?: boolean
     requestStartTime?: number
-    mcpAnalyticsProvider?: McpAnalyticsProvider
-    mcpAnalyticsFlagKey?: string
-    mcpAnalyticsFlagEnabled?: boolean
-    mcpAnalyticsFlagErrorName?: string
-    mcpAnalyticsFlagErrorMessage?: string
-    posthogMcpAnalyticsInitAction?: PostHogMcpAnalyticsInitResult['action']
-    posthogMcpAnalyticsInitReason?: string
-    posthogMcpAnalyticsInitErrorName?: string
-    posthogMcpAnalyticsInitErrorMessage?: string
+    mcpAnalyticsInitAction?: McpAnalyticsInitResult['action']
+    mcpAnalyticsInitReason?: string
+    mcpAnalyticsInitErrorName?: string
+    mcpAnalyticsInitErrorMessage?: string
 }
 
 export class MCP extends McpAgent<Env> {
@@ -540,7 +528,6 @@ export class MCP extends McpAgent<Env> {
         const flagPromise = this.resolveVersionFlag()
         const toolFlagsPromise = this.resolveToolFeatureFlags(clientVersion)
         const singleExecPromise = this.resolveSingleExecFlag()
-        const posthogMcpAnalyticsFlagPromise = this.resolvePostHogMcpAnalyticsFlag()
 
         // Seed cache with header-provided IDs before any fetches
         if (organizationId) {
@@ -569,15 +556,13 @@ export class MCP extends McpAgent<Env> {
             await context.stateManager.setDefaultOrganizationAndProject()
         }
 
-        const [flagVersion, toolFeatureFlags, singleExecFlagOn, posthogMcpAnalyticsFlag] = await Promise.all([
+        const [flagVersion, toolFeatureFlags, singleExecFlagOn, _apiKey] = await Promise.all([
             flagPromise,
             toolFlagsPromise,
             singleExecPromise,
-            posthogMcpAnalyticsFlagPromise,
             // Trigger OAuth introspection so the OAuth client name is cached before the useSingleExec decision below
             context.stateManager.getApiKey(),
         ])
-        const isPosthogMCPAnalyticsEnabled = posthogMcpAnalyticsFlag.enabled
 
         const oauthClientName = (await this.cache.get('clientName')) || undefined
 
@@ -768,61 +753,42 @@ export class MCP extends McpAgent<Env> {
             getMcpConversationId: async () => this.requestProperties.mcpConversationId,
         }
 
-        Object.assign(this.requestProperties, {
-            mcpAnalyticsProvider: isPosthogMCPAnalyticsEnabled ? 'posthog_mcp_analytics' : 'mcpcat',
-            mcpAnalyticsFlagKey: POSTHOG_MCP_ANALYTICS_FLAG,
-            mcpAnalyticsFlagEnabled: isPosthogMCPAnalyticsEnabled,
-            ...(posthogMcpAnalyticsFlag.errorName
-                ? { mcpAnalyticsFlagErrorName: posthogMcpAnalyticsFlag.errorName }
-                : {}),
-            ...(posthogMcpAnalyticsFlag.errorMessage
-                ? { mcpAnalyticsFlagErrorMessage: posthogMcpAnalyticsFlag.errorMessage }
-                : {}),
+        // In single-exec mode every event's `$mcp_tool_name` is `exec`, so the
+        // SDK's `$mcp_tool_description` would be the dispatcher's static text on
+        // every call. Resolve the inner tool the agent was actually invoking
+        // from the command and surface its name + description as
+        // `$mcp_exec_tool_call_name` / `$mcp_exec_tool_call_description`.
+        const resolveExecInnerToolCall = useSingleExec ? createExecInnerToolCallResolver(allTools) : undefined
+
+        // In single-exec mode the SDK's $mcp_listed_tool_names collapses to
+        // just `exec`, so we can't compute "advertised but never called"
+        // (zombie tools) from SDK data alone. Pass the inner-tool catalog
+        // here so analytics can attach it as $mcp_exec_inner_tool_names on
+        // mcp_tools_list events. Dashboards can then diff it against
+        // $mcp_exec_tool_call_name from mcp_tool_call.
+        const execInnerToolNames = useSingleExec ? allTools.map((t) => t.name) : undefined
+
+        const initResult = await initMcpAnalytics(this.server, mcpAnalyticsIdentity, {
+            contextEnabled: true,
+            resolveExecInnerToolCall,
+            execInnerToolNames,
+            // `get_more_tools` only earns its keep outside single-exec mode — there
+            // it lets the model report a gap in our discrete tool catalog. In
+            // single-exec mode the wrapper handles every call, so the missing-tool
+            // signal has nothing to map to and the extra slot is pure noise.
+            reportMissingEnabled: !useSingleExec,
         })
 
-        // @posthog/mcp is based on the MCP Cat wrapper and keeps tool tracing,
-        // AI spans, context capture, and get_more_tools for flagged sessions.
-        // Avoid initializing both wrappers because both patch tool handlers and
-        // would double-capture every call.
-        // `get_more_tools` only earns its keep outside single-exec mode — there
-        // it lets the model report a gap in our discrete tool catalog. In
-        // single-exec mode the wrapper handles every call, so the missing-tool
-        // signal has nothing to map to and the extra slot is pure noise.
-        if (isPosthogMCPAnalyticsEnabled) {
-            // In single-exec mode every event's `$mcp_tool_name` is `exec`, so the
-            // SDK's `$mcp_tool_description` would be the dispatcher's static text on
-            // every call. Resolve the inner tool the agent was actually invoking
-            // from the command and surface its name + description as
-            // `$mcp_exec_tool_call_name` / `$mcp_exec_tool_call_description`.
-            const resolveExecInnerToolCall = useSingleExec ? createExecInnerToolCallResolver(allTools) : undefined
-
-            // In single-exec mode the SDK's $mcp_listed_tool_names collapses to
-            // just `exec`, so we can't compute "advertised but never called"
-            // (zombie tools) from SDK data alone. Pass the inner-tool catalog
-            // here so analytics can attach it as $mcp_exec_inner_tool_names on
-            // mcp_tools_list events. Dashboards can then diff it against
-            // $mcp_exec_tool_call_name from mcp_tool_call.
-            const execInnerToolNames = useSingleExec ? allTools.map((t) => t.name) : undefined
-
-            const initResult = await initPostHogMcpAnalytics(this.server, mcpAnalyticsIdentity, {
-                contextEnabled: true,
-                reportMissingEnabled: !useSingleExec,
-                resolveExecInnerToolCall,
-                execInnerToolNames,
-            })
-            Object.assign(this.requestProperties, {
-                posthogMcpAnalyticsInitAction: initResult.action,
-                ...(initResult.action === 'skipped' ? { posthogMcpAnalyticsInitReason: initResult.reason } : {}),
-                ...(initResult.action === 'failed'
-                    ? {
-                          posthogMcpAnalyticsInitErrorName: initResult.errorName,
-                          posthogMcpAnalyticsInitErrorMessage: initResult.errorMessage,
-                      }
-                    : {}),
-            })
-        } else {
-            await initMcpCatObservability(this.server, mcpAnalyticsIdentity)
-        }
+        Object.assign(this.requestProperties, {
+            mcpAnalyticsInitAction: initResult.action,
+            ...(initResult.action === 'skipped' ? { mcpAnalyticsInitReason: initResult.reason } : {}),
+            ...(initResult.action === 'failed'
+                ? {
+                      mcpAnalyticsInitErrorName: initResult.errorName,
+                      mcpAnalyticsInitErrorMessage: initResult.errorMessage,
+                  }
+                : {}),
+        } as RequestProperties)
 
         const initDurationMs = this.requestProperties.requestStartTime
             ? Date.now() - this.requestProperties.requestStartTime
@@ -903,19 +869,6 @@ export class MCP extends McpAgent<Env> {
             return !!(await isFeatureFlagEnabled('mcp-single-exec-tool', distinctId))
         } catch {
             return false
-        }
-    }
-
-    private async resolvePostHogMcpAnalyticsFlag(): Promise<PostHogMcpAnalyticsFlagResult> {
-        try {
-            const distinctId = await this.getDistinctId()
-            return { enabled: !!(await isFeatureFlagEnabled(POSTHOG_MCP_ANALYTICS_FLAG, distinctId)) }
-        } catch (error) {
-            return {
-                enabled: false,
-                errorName: error instanceof Error ? error.name : 'UnknownError',
-                errorMessage: error instanceof Error ? error.message : String(error),
-            }
         }
     }
