+++ services/mcp/src/tools/types.ts
@@ -2,8 +2,8 @@ import type { z } from 'zod'
 
 import type { ApiClient, GroupType } from '@/api/client'
 import type { Schemas } from '@/api/generated'
-import type { AnalyticsEvent } from '@/lib/analytics'
 import type { ScopedCache } from '@/lib/cache/ScopedCache'
+import type { AnalyticsEvent } from '@/lib/posthog/analytics'
 import type { SessionManager } from '@/lib/SessionManager'
 import type { StateManager } from '@/lib/StateManager'
 import type { PrefixedString } from '@/lib/types'
